u32 delay
(cherry picked from commit e44be8771be749d3efbd8e0f31081dc8f7c187d6)

+++ src/notification.rs
@@ -222,7 +222,7 @@ impl Notifications {
 
         if last_notification.is_some_and(|last_notification| {
             now - last_notification
-                < TimeDelta::milliseconds(config.delay.unwrap_or(500))
+                < TimeDelta::milliseconds(config.delay.unwrap_or(500) as i64)
         }) {
             return;
         }
